[TT-13175] Refactor apply policies test (#6585)
### **User description**
<!-- Provide a general summary of your changes in the Title above -->

## Description
refactor apply policies tests to internal/policy

## Related Issue
https://tyktech.atlassian.net/browse/TT-13175

## Motivation and Context

<!-- Why is this change required? What problem does it solve? -->

## How This Has Been Tested

<!-- Please describe in detail how you tested your changes -->
<!-- Include details of your testing environment, and the tests -->
<!-- you ran to see how your change affects other areas of the code,
etc. -->
<!-- This information is helpful for reviewers and QA. -->

## Screenshots (if appropriate)

## Types of changes

<!-- What types of changes does your code introduce? Put an `x` in all
the boxes that apply: -->

- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing
functionality to change)
- [x] Refactoring or add test (improvements in base code or adds test
coverage to functionality)

## Checklist

<!-- Go over all the following points, and put an `x` in all the boxes
that apply -->
<!-- If there are no documentation updates required, mark the item as
checked. -->
<!-- Raise up any additional concerns not covered by the checklist. -->

- [ ] I ensured that the documentation is up to date
- [ ] I explained why this PR updates go.mod in detail with reasoning
why it's required
- [ ] I would like a code coverage CI quality gate exception and have
explained why


___

### **PR Type**
Tests


___

### **Description**
- Refactored apply policies test cases by moving them from
`gateway/policy_test.go` to `internal/policy/apply_test.go`.
- Removed old test functions `TestApplyPolicies` and
`BenchmarkApplyPolicies` from `gateway/policy_test.go`.
- Added new test functions `TestService_Apply` and
`BenchmarkService_Apply` in `internal/policy/apply_test.go`.
- Introduced new policy entries in
`internal/policy/testdata/policies.json` for enhanced test coverage.


___



### **Changes walkthrough** 📝
<table><thead><tr><th></th><th align="left">Relevant
files</th></tr></thead><tbody><tr><td><strong>Tests</strong></td><td><table>
<tr>
  <td>
    <details>
<summary><strong>policy_test.go</strong><dd><code>Remove apply policies
test cases from gateway</code>&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp;
&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; </dd></summary>
<hr>

gateway/policy_test.go

<li>Removed extensive test cases related to policy application.<br> <li>
Deleted the <code>TestApplyPolicies</code> and
<code>BenchmarkApplyPolicies</code> functions.<br> <li> Removed imports
related to deleted test cases.


</details>


  </td>
<td><a
href="https://github.com/TykTechnologies/tyk/pull/6585/files#diff-40d701767204255c38c7dd64939d6bb8df621640c4bddfe5f56080380476a18a">+0/-1088</a></td>

</tr>                    

<tr>
  <td>
    <details>
<summary><strong>apply_test.go</strong><dd><code>Add apply policies test
cases to internal/policy</code>&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp;
&nbsp; &nbsp; &nbsp; </dd></summary>
<hr>

internal/policy/apply_test.go

<li>Added new test cases for policy application.<br> <li> Introduced
<code>TestService_Apply</code> and <code>BenchmarkService_Apply</code>
functions.<br> <li> Added helper function
<code>testPrepareApplyPolicies</code>.<br> <li> Imported necessary
packages for new tests.


</details>


  </td>
<td><a
href="https://github.com/TykTechnologies/tyk/pull/6585/files#diff-5af7e299a6b0ce11e22f8aa4a01854b1151f4b54dccc68f0cd1cbedee5aed7c8">+1077/-0</a></td>

</tr>                    

<tr>
  <td>
    <details>
<summary><strong>policies.json</strong><dd><code>Add new policy entries
for testing</code>&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp;
&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp;
&nbsp; &nbsp; &nbsp; &nbsp; &nbsp; &nbsp; </dd></summary>
<hr>

internal/policy/testdata/policies.json

<li>Added new policy entries for testing.<br> <li> Introduced
<code>acl_with_allowed_url</code> and <code>rate_limit</code> policies.


</details>


  </td>
<td><a
href="https://github.com/TykTechnologies/tyk/pull/6585/files#diff-098134f4708ce69a2d65ed25b4e81b76ad30a54ace915d8c54e4630984038804">+37/-0</a>&nbsp;
&nbsp; </td>

</tr>                    
</table></td></tr></tr></tbody></table>

___

> 💡 **PR-Agent usage**: Comment `/help "your question"` on any pull
request to receive relevant information
